impl: Added contracts already deployed check

+++ scripts/deploy_docker.sh
@@ -1,14 +1,17 @@
 #!/bin/bash
 
+if [ -f "/ocean-contracts/ccache/deployed" ]
+then
+    echo "Contracts already deployed"
+    exit 0
+fi
+
+
 if [ "${DEPLOY_CONTRACTS}" = "true" ]
 then
     echo "Deploying Contracts..."
-    rm /ocean-contracts/artifacts/ready
+    rm -f /ocean-contracts/artifacts/ready
 
-    npx hardhat clean
-    npx hardhat compile
-    #remove unneeded debug artifacts
-    find /ocean-contracts/artifacts/* -name "*.dbg.json" -type f -delete
     #copy address.json
     if [ -e /ocean-contracts/addresses/address.json ]
         then cp -u /ocean-contracts/addresses/address.json /ocean-contracts/artifacts/
@@ -16,4 +19,5 @@ then
     echo "Starting deployment process..."
     node /ocean-contracts/scripts/deploy-contracts.js
     touch /ocean-contracts/artifacts/ready
+    touch /ocean-contracts/ccache/deployed
 fi
